3NF Normalization and Interface
This project has been developed using Python and MySQL. It is a database design for a game. To establish entity relationships in the database, 3NF normalization has been applied.

In this project, Python programming language is used to process game data and perform database operations through various functions and classes. Tables are created using MySQL database to represent different components and relationships of the game.

3NF (Third Normal Form) is a level of normalization in database design. This normalization level is used to organize the database data effectively and accurately represent the relationships between data. 3NF normalization reduces unnecessary data repetition in tables and enhances data consistency.

This project demonstrates a robust and well-organized database design. The combination of Python and MySQL provides a powerful solution for efficiently managing game data.